Slight edit to test for Box.unbox()
